changed warning message, moved deprecated function to api_versioning
