+++ usaspending_api/common/api_versioning.py
@@ -1,10 +1,13 @@
 from django.utils.decorators import method_decorator
 from usaspending_api.common.exceptions import InvalidParameterException
 from usaspending_api.awards.v2.filters.filter_helpers import transform_keyword
+import logging
+
 
 API_TRANSFORM_FUNCTIONS = [
     transform_keyword,
 ]
+logger = logging.getLogger('console')
 
 
 def api_transformations(api_version, function_list):
@@ -26,3 +29,14 @@ def api_transformations(api_version, function_list):
         ClassBasedView.post = method_decorator(view_func)(ClassBasedView.post)
         return ClassBasedView
     return class_based_decorator
+
+
+def deprecated(function):
+    """Add deprecation warning to endpoint"""
+    def wrap(request, *args, **kwargs):
+        response = function(request, *args, **kwargs)
+        logger.warning('Endpoint "{}" is deprecated. Please move to v2 endpoints.'.format(request.path))
+        response["X-API-Warn"] = "WARNING! You are using a deprecated version of the API"
+        response["X-API-Warn"] += " which may be unstable and will be removed in the future."
+        return response
+    return wrap
